# Fix for empty flamegraph on linux