@david-martin 2016-10-17
Note to the reviewer, the diff looks more complex than the actual change.
It is essentially wrapping existing code with:
```javascript
process.on('exit', function() {
  // existing code ...
});
```