@david-martin 2016-10-18
@lzuccarelli Would you be able to try this locally?
If you clone & checkout this branch,
then `npm link .` globally for the `0x` cli