@luigizuccarelli 2016-10-18
@david-martin  - Installed and tested this locally - works like a charm