@davidmarkclements 2016-10-18
published to npm as v2.2.12